+++ Changelog.md
@@ -4,6 +4,9 @@
    set the received message rate on the subscriber side.
     * [Pull request 174](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/174)
 
+1. Added ign service --req <args ...> for requesting services using the command line.
+    * [Pull request 172](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/172)
+
 1. Do not allow to advertise a topic that is currently advertised on the same node.
    See [issue #54] (https://bitbucket.org/ignitionrobotics/ign-transport/issues/54)
     * [Pull request 169](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/169)